Revert "chore: folder structure refactor"
This reverts commit d9f3db89bb8085ea1a16f4be0bd6d605ea20dca4.
