+++ client/src/features/event-editor/EventEditor.jsx
@@ -2,9 +2,9 @@ import { useCallback, useContext, useEffect, useState } from 'react';
 import { Button, Select, Switch } from '@chakra-ui/react';
 import { IoBan } from '@react-icons/all-files/io5/IoBan';
 import { editorEventId } from 'common/atoms/LocalEventSettings';
-import ColourInput from 'common/components/input/colour-input/ColourInput';
-import TextInput from 'common/components/input/text-input/TextInput';
-import TimeInput from 'common/components/input/time-input/TimeInput';
+import ColourInput from 'common/components/input/ColourInput';
+import TextInput from 'common/components/input/TextInput';
+import TimeInput from 'common/components/input/TimeInput';
 import { LoggingContext } from 'common/context/LoggingContext';
 import { useEventAction } from 'common/hooks/useEventAction';
 import { millisToMinutes } from 'common/utils/dateConfig';
