+++ client/src/features/rundown/delay-block/DelayBlock.tsx
@@ -3,7 +3,7 @@ import { Draggable } from 'react-beautiful-dnd';
 import { Button, HStack } from '@chakra-ui/react';
 import { IoCheckmark } from '@react-icons/all-files/io5/IoCheckmark';
 import { IoReorderTwo } from '@react-icons/all-files/io5/IoReorderTwo';
-import DelayInput from 'common/components/input/delay-input/DelayInput';
+import DelayInput from 'common/components/input/DelayInput';
 import { useEventAction } from 'common/hooks/useEventAction';
 import { millisToMinutes } from 'common/utils/dateConfig';
